[Issue #73] @Builder.Extension
If a @Builder.Extension is a required or optional extension, is no longer decided by looking at the assigned fields, since this approach didn't work flawlessly in the Eclipse IDE. Now the method name classifies the extension type like this:

method name: <PREFIX>fieldName1<And>fieldName2...()

If both fieldName1 and fieldName2 represent optional fields than the extension is optional. If both fieldName1 and fieldName2 represent required fields than the extension is required. Otherwise the handler errors out since the extension is falsely configured.